fix: update existing qrcode test
